rename knowledge_graph folder
